--- app/views/event_signup_mailer/new_signup.html.haml
@@ -1,5 +1,0 @@
-%h1 New signup for #{run_description @signup.run}
-
-%p= signup_description(@signup, "has signed up for", html: true)
-
-= simple_format "As of this signup, this event run has:\n#{@signup_count_presenter.signups_description}"
