--- app/views/event_signup_mailer/new_signup.text.erb
@@ -1,6 +1,0 @@
-New signup for <%= run_description @signup.run %>
-
-<%= signup_description(@signup, "has signed up for", html: false) %>
-
-As of this signup, this event run has:
-<%= @signup_count_presenter.signups_description %>
