+++ cms_content_sets/standard/notification_templates/signups/new_signup.liquid
@@ -1,0 +1,12 @@
+---
+subject: "[{{ convention.name }}: {{ signup.event.title }}] Signup: {{ signup.user_con_profile.name }}"
+---
+<h1>New signup for {{ run_description }}</h1>
+<p>
+  {{ signup_description }}
+</p>
+<p>
+  As of this signup, this event run has:
+  <br>
+  {{ all_signups_description | newline_to_br }}
+</p>
